Add a wrapper for the RIL
On Android 11, the property 'gsm.version.baseband' is too long on dual
sim phones. This is caused by the limited lenght of Android properties
(91 characters) and by the fact that Android now set this property as
'basebandSIM1,basebandSIM2'.

To avoid this, I added this libril-wrapper that make sure to get the
lenght of the baseband version under 45 characters when requested to the
HAL.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>