Add KeyHandler for the tri-state key
The OnePlus phones have a tri-state key to control volume (silence,
vibration, normal ring).
This is a bit complex as Android does not provide controls over
unknown buttons, and in fact LineageOS adds support for this on
frameworks/base.

We instead act differenctly: we add an accessibility service (that is
able to enable itself at first boot) to control the button when the
screen in on.
To control the buttons with screen off, we setup those keys as 'WAKE' and we
modify the kernel to send the events twice:
- the user uses the slider,
- the kernel send two events,
- at first event, Android exits from deep sleep,
- at second event, KeyHandler change the volume status and ask the
device to get into deep sleep again (turning off the screen).

Due to some timing luck, the screen does not actually turn on (the user
will not notice anything).

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>