KeyHandler: set the correct mode at boot
Read the Tri Key position at boot from sysfs and set the mode (normal,
vibrate or silent) accordingly.
This also fixes an issue: at boot, Android never sets the Media Volume
to mute for some reason. With this, we make sure to be silent in case of
a reboot in silent mode.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>