KeyHandler: switch from DevicePolicy to PowerManager to turn off the screen
When the screen is turned off via DevicePolicyManager, the device is
also locked. In this case, if the fingerprint is used to unlock the
device, a message 'device locked by admin' is shown on screen and the
PIN is requested.
To avoid this issue, that makes the fingerprint useless, turn off the
screen using the PowerManager API.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>