config: keylayout: add tri-state slider management
This is an hack needed for the KeyHandler app.
To use the slider with the screen off, we have to be able to turn on the
device when the slider changes: to do this, map those 'buttons' as DPAD
buttos because Android can wake up if a DPAD is used.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>