sepolicy: add label for the Tri State slider status
Allow system applications to access the slider status. We use this to
set the correct mode (normal, vibrate, silent) at boot, as the driver
sends interrupts that we cannot get with the phone off obviously.

It also helps us setting to silent the media volume at boot (Android
does not seem happy to remember this after a reboot).

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>