Refactoring and new features for the old KeyHandler app: now Settings Extra
As the KeyHandler app has started to do more things, we rename this as
'Settings Extra'. It now has an interface where the user can toggle some
options, like:
- swap back and recent capacitive buttons,
- change display color gamut (sRGB, DCI-P3),
- enable or disable the display DC dimming.

It is also using the Device Protected storage to store the preferences
now, so that it has available them earlier at boot.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>