SettingsExtra: add NFC tile service
AOSP is somehow using a broken NFC tile: implement a working tile using
the Settings Extra app and Android APIs.

The KeyHandler service is used to dispatch those events to the NfcTile
class: this is because that service is always running, and it does not
make sense to add a new service just for this. IN the future, the
KeyHandler service should change name to be more generic.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>