SettingsExtra: change KeyHandler to MainService
This service is now doing more than just managing key events. For
example, it listens also for NFC events.
Rename this to MainService, and enable at every boot: it is now
important and we don't want it disabled. Also, we changed name to it, so
in case of an upgrade we force this service to be enabled again.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>