SettingsExtra: fix NFC tile startup
So, NFC is broken on Android. The NFC service starts too late in the
boot process causing errors when any application is trying to use before
that.
To avoid this issue, we start a new thread that will update the NFC tile
at boot: the NFC tile class now check for the NFC service availability
for 30 seconds with a series of bad try and catch.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>