SettingsExtra: add double tap to wake mechanism while Always On Display is enabled
Android does not wake the device from Always On Display mode if both
options are enabled.
As we have the MainService listening for key events, enabled the F4 key
events from /proc/touchpanel/tp_f4 only when the screen is going off and
AOD is enabled. In that case, wait for a double tap to exit from AOD and
get into the lockscreen.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>