SettingsExtra: add battery level to the bluetooth icon on status bar
Android does not show any information on the status bar regarding the
Bluetooth battery level: remove the AOSP status bar icon add add the
LineageOS icon, that has an indicator with the battery level.
Resources from here:
https://review.lineageos.org/c/LineageOS/android_frameworks_base/+/288694

To avoid editing the SystemUI, we remove at boot the default bluetooth
icon and create a new one that will be present only when a bluetooth
device (with a battery) is connected. This new icon will change in case of a battery
status change on the bluetooth device.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>