SettingsExtra: big, big refactoring and introduction of Dynamic Modem
The MainService was becoming to big, so create three different classes
- BluetoothBatteryIcon
- CallRecording
- NfcStatusMonitor

This allows to remove a lot of code from the MainService and move that
here: we now just start the services from the MainService.

Also, as we are now using Android Studio, everything is indented, only
the useful things are included and stuff like that.

While we were there, we added also the new Dynamic Modem. It basically
allows to select the network mode (GSM, LTE, UMTS, etc..) based on the
wifi connection: the network mode will automatically switch based on the
wifi connection. The settings are configurable from the SettingsExtra
UI.

Both the refactoring and the Dynamic Modem may still have issues that
will be solved in the future.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>