SettingsExtra: add UserBootCompletedReceiver class
This class is a receiver for the BOOT_COMPLETED intent. It is used to
manage settings or start activities that does not need to be started at
LOCKED_BOOT_COMPLETED.

At the moment, it is used to set the initial states for the NFC and
Adaptive Brightness tiles because there are issues when those tiles are
updated too early during the boot process.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>