SettingsExtra: improve mechanism to manage screen off slider events
When the screen is off, we do some tricks to not turn on the screen in
case of slider events (mute, vibration or silent mode).
The 'old' mechanism was based on the timings of the key events, but now
we look at the SCREEN_ON and SCREEN_OFF events: this seems more stable
and should avoid potential issues (screen turning on randomly) that I
saw durimg my tests.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>