config: BoardConfig: set ALLOW_MISSING_DEPENDENCIES
We are not building some LineageOS applications (e.g. PocketMode,
LiveDisplay) that are in the build system and requires libraries that we
don't have, however we get errors anyway as those are actually missing
dependencies. Ignore that, but we should set this property to false
as soon as we can.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>