Add the tri-state-key_daemon
This is a native application that is listening for uevent events. When
an event from the tri state slider is detected, it launches a broadcast
so that the Android applications can detect it.
We are going to use this so that SettingsExtra can detect the slider
changes and set the volume consequently, even with screen off.

It allows us to remove a lot of tricks also in the kernel.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>