SettingsExtra: refactor again the mechanism to manage screen off slider events [2]
Change the tri-state-key_daemon mechanism: now, it is looking at
accesses to the /dev/input/event# device (only the one of the slider) to launch
the broadcast. This is not ideal, however I've noticed that with the
previous implementation sometimes the broadcast was not launched - in
particular, after the user unlocked the screen the first time after the
boot, the broadcast was not launched for several seconds.

We are using the inotify kernel API so this should be pretty efficient.

Add also the SELinux rules to look into sysfs and find the corrent input
device, and change SettingsExtra so that now the feedback (vibration)
starts before the actual audio settings are changed: this gives a faster
feedback to the user.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>