Refactor again the mechanism to manage screen off slider events [2]
Improve again the mechanism to manage the slider events: now the
SettingsExtra app does nothing, and everything is managed by the native
tri-state-key_daemon.
This is caused by the fact that with the previous implementation sometimes
the broadcast was not launched - in particular, after the user unlocked
the screen the first time after the boot, the broadcast was not launched
for several seconds (up to one minute delay!).

Update the SELinux rules consequently, and start the vibration feedback
before the actual audio settings are changed: this gives a faster
feedback to the user.

Note that now we are using 'service call' to set the audio policy, and
these calls can change when Google decide it (they are not stable as the
public java API).

Last thing: we cannot write on the vibrator sysfs node due to SELinux,
so we leave this job to init.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>