Add custom Light HAL
This HAL is managed via Android properties and it bypass the Android
HAL. It allows for full customization from any Android apps instead of
relying on the Android framework for the LED management (that is very
limited).
SettingsExtra will provide the user a LED Manager.

Also add the needed sepolicy rules and include the HAL binaries in the
OS.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>