SettingsExtra: add LED Manager
Considera this a BETA feature for the moment.

LED Manager provides an interface to manage the LED on the OnePlus 5.
It allows for fully configuration of the LED colors including blinking
and string matching on the notification titles.

When enabled from SettingsExtra, the default Android HAL is disabled.
When enabled, the Android LED API will not work anymore: the colors
indicated in any application will not work: the only way to turn on the
LED is to use SettingsExtra.
Note that no error will be generated when a third party will try to turn
on the LED: the LED will simply not work.

The Led colors, blinking effect and filters are configurable, but there is no
UI available.
The configuration can be injected using a text file, for example:

com.example.myapplication,,00FFFF,
com.example.myapplication,display,FFFF00,blink,500,5000
com.example.myapplication,test,0000FF,blink,500,500
android,debugging,FF00FF,blink,500,5000

The file should contains only lines with that structure, it should be named
SettingsExtraLedConfiguration.conf, and it should be push as:
adb push SettingsExtraLedConfiguration.conf /sdcard/

The rules are:
- color are in hexadecimal, and they can be generated here (any
website will work fine): https://htmlcolorcodes.com/
- the structure of every line is:
application package name, string, color, blink, milliseconds ON, milliseconds OFF

where:
- application package name: the application package name, as com.whatsapp
for Whatsapp or com.android.settings for the Android Settings. You can use
simple 'android' for the SystemUI notification (for example for the USB
debugging notification)
- string: a word or a phrase to match the title of the notification. For
example, setting as application package name 'com.whatsapp' and as
string 'Name Surnane', this rules will only apply to notifications
including the text 'Name Surname' in the notification title. Useful to
filter for particular contact, for example. It is case insensitive. Leave this
field empty to assign the same color to all the notification from the
indicated application.
- color: the hexadecimal color of the led
- bling: use the work 'blink' if you want the led to blink. Leave this
empty, and not include the milliseconds parameters if you want the led
to be always on
- milliseconds ON: integer value to indicate how much the led should stay on
- milliseconds OFF: integer value to indicate how much the led should stay off

The led is hardcoded to be yellow while the phone is charging and green when
the phone is charging but with 100% battery.
The priorities  are:
- the charging led colors have the lowest priority -> if a notification with
a color is present, the charging status will not be shown
- if multiple notifications are present, the latest notification color is shown
- when one notification goes away, the color of the latest notification
appeared will be shown
- the rule with a string indicated has priority over a rule for the same app
but without matching string.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>