config: system.prop: disable package parsing cache
Apparently, Android has a 'package parsing cache' in the
'data/system/package_cache' directory. This has caused a lot of issues*
to me, so I just disable it here.

* For example, it broke the 'Search' feature on the Settings app
somehow. It also broke phone calls when we moved the telephony
packages from 'system' to 'system_ext', as this cache was still
reporting the packages as available in system.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>