SettingsExtra: turn on the device when an alarm has to be triggered
If the device is completely turned off, but we have an alarm (set by an
Android app) that has to be triggered, we should turn on the device to
show it to the user.
Qualcomm thought about this, but of course with a closed source application
(PowerOffAlarm.apk). This application can communicate with the RTC to allow
the behaviour described above, and it receives broadcast as
'org.codeaurora.poweroffalarm.action.SET_ALARM'.

So, listen for new alarms added and send this broadcast when needed from
SettingsExtra. The PowerOffAlarm app will receive the broadcast and set
the correct wake up time on the RTC.

Before sending the intent, we subtract 90 seconds from the time decided
by the user so that the device will power on 90 seconds before that and
it will actually show the alarm to the user. Without this trick, Android
will only show the 'missed alarms' alert without actually doing anything.
There is no change to the actual alarm time.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>