SettingsExtra: fix power on device in case of multiple alarms
If multiple alarms are set, we need to make sure to only have one alarm
set on the RTC. This is because Android does not provide API to access
the alarms list.
For this reason, before setting a new alarm, remove the previous one
from the RTC.

The Android alarms are not touched by this.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>