SettingsExtra: use PreciseCallState for CallManager
This allows us to precisely detect call states. In particular,
CallManager now knows exactly when a call actually starts - meaning that
we picked up or the other side picked up.
Before this change, we started recording a call or vibrating immediatly
after the user dialed the number. Now, we wait for the other side to
pick up.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>