SettingsExtra: add support for gestures and on-screen navigation bar
Use the 'qemu.hw.mainkeys' to support gestures and/or set the on-screen
navigation bar. An option to enable on-screen navigation bar is added to
SettingsExtra:
- gestures enabled -> qemu.hw.mainkeys = 0;
- gestures disabled and navigation bar enabled -> qemu.hw.mainkeys = 0;
- gestures disabled and navigation bar disabled -> qemu.hw.mainkeys = 1;

When gestures are enabled, qemu.hw.mainkeys is forced to 0.

A toast asking for a reboot will be shown when needed, as to change the
property a reboot is needed.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>