build: task: introduce the 'otapackage_norecovery' target
This is basically a copy of the 'otapackage' target, with the only
difference that it changes the 'install-recovery.sh' script so that the
installed recovery is not updated or overwritten.
This is useful in case the user is using something like a TWRP recovery.

To do this without modifying AOSP and without copying too much code, we
create the target files as usual (including the 'install-recovery.sh'
script) but then - before actually creating the OTA zip - we edit the
script to not do anything to the recovery itself.

Note that we cannot remove the script, as the python scripts that
generate the recovery OTA check for its presence.
We need also to modify both the script in the target zip archive and the
script in the 'vendor.img' image (also in the target zip archive). For this
reason, this process requires to zip/unzip some archives and run some
AOSP scripts, and that can takes a bit based on the disk speed on the
host pc.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>