Add possibility to use QPST on Android 13
QPST/EFS Explorer/EfsTools are (Windows) tools used to modify the EFS data:
this is useful for example to enable VoLTE/VoWifi on phones without direct
support from the carrier.

The 'diag' kernel driver is used to manage the communication between
phone and host pc.

On custom rom, two things are missing to get this working:
- correct SELinux permissions
- the init.rc configfs configuration.

Copy the second one from OxygenOS (Android 9), and add the needed rule
to get it working.

The SELinux rule is present also on Pixel 3:
https://android.googlesource.com/device/google/crosshatch-sepolicy/+/refs/heads/android12-release/vendor/qcom/common/kernel.te#1

This voluntarily requires root to work, as it may damage the phone
(breaking phone calls/3G/4G functionalities, without possibility to fix
them).

From an adb root shell:
:/ # setprop sys.usb.config diag,serial_cdev,serial_tty,rmnet_ipa,mass_storage,adb

Automatically, a new serial port will be exported on the host pc if the
correct driver is installed (running Windows - the tools mentioned
above only work on Windows).

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>